adding assertions
