+++ pyproject.toml
@@ -52,3 +52,6 @@ cfn-flip = "==1.2.3"
 networkx = "^2.5"
 troposphere = "^2.6.3"
 awacs = "^1.0.2"
+jmespath = "^0.10.0"
+deepdiff = "^5.3.0"
+deepmerge = "^0.2.1"
