+++ servicecatalog_puppet/manifest_utils.py
@@ -106,9 +106,9 @@ def expand_manifest(manifest, client):
         account_id = account.get("account_id")
         if account.get("append") or account.get("overwrite"):
             if (
-                account.get("default_region")
-                or account.get("regions_enabled")
-                or account.get("tags")
+                    account.get("default_region")
+                    or account.get("regions_enabled")
+                    or account.get("tags")
             ):
                 raise Exception(
                     f"{account_id}: If using append or overwrite you cannot set default_region, regions_enabled or tags"
@@ -146,7 +146,7 @@ def expand_manifest(manifest, client):
 
     for launch_name, launch_details in new_manifest.get(constants.LAUNCHES, {}).items():
         for parameter_name, parameter_details in launch_details.get(
-            "parameters", {}
+                "parameters", {}
         ).items():
             if parameter_details.get("macro"):
                 macro_to_run = macros.get(parameter_details.get("macro").get("method"))
@@ -164,6 +164,8 @@ def rewrite_depends_on(manifest):
         (constants.LAUNCH, constants.LAUNCHES),
         (constants.SPOKE_LOCAL_PORTFOLIO, constants.SPOKE_LOCAL_PORTFOLIOS),
         (constants.LAMBDA_INVOCATION, constants.LAMBDA_INVOCATIONS),
+        (constants.CODE_BUILD_RUN, constants.CODE_BUILD_RUNS),
+        (constants.ASSERTION, constants.ASSERTIONS),
         (constants.ACTION, constants.ACTIONS),
     ]
     for section_item_name, section_name in sections:
@@ -178,7 +180,6 @@ def rewrite_depends_on(manifest):
                         details["depends_on"][i][constants.AFFINITY] = details[
                             "depends_on"
                         ][i]["type"]
-        print(manifest.get(section_name))
     return manifest
 
 
@@ -265,30 +266,72 @@ def get_from_dict(d, path):
 
 
 class Manifest(dict):
-    def get_tasks_for_launch(self, puppet_account_id, launch_name):
+    def get_tasks_for(self, puppet_account_id, section_name, item_name):
         accounts = self.get(constants.ACCOUNTS)
-        section = self.get(constants.LAUNCHES)
+        section = self.get(section_name)
         provisioning_tasks = list()
-        item = section[launch_name]
-        deploy_to = "deploy_to"
+        item = section[item_name]
 
-        common_parameters = dict(
-            puppet_account_id=puppet_account_id,
-            launch_name=launch_name,
-            launch_parameters=item.get("parameters", {}),
-            manifest_parameters=self.get("parameters", {}),
-            ssm_param_outputs=item.get("outputs", {}).get("ssm", []),
-        )
-        for p in [
-            "portfolio",
-            "product",
-            "version",
-            "execution",
-            "requested_priority",
-            "execution",
-        ]:
-            if item.get(p):
-                common_parameters[p] = item.get(p)
+        deploy_to = {
+            "launches": "deploy_to",
+            "spoke-local-portfolios": "share_with",
+            "lambda-invocations": "invoke_for",
+            "code-build-runs": "run_for",
+            "assertions": "assert_for",
+        }.get(section_name)
+
+        if section_name == constants.SPOKE_LOCAL_PORTFOLIOS and item.get(deploy_to) is None:
+            deploy_to = "deploy_to"
+
+        common_parameters = {
+            "launches": dict(
+                puppet_account_id=puppet_account_id,
+                launch_name=item_name,
+                launch_parameters=item.get("parameters", {}),
+                manifest_parameters=self.get("parameters", {}),
+                ssm_param_outputs=item.get("outputs", {}).get("ssm", []),
+                portfolio=item.get("portfolio"),
+                product=item.get("product"),
+                version=item.get("version"),
+                execution=item.get("execution"),
+                requested_priority=item.get("requested_priority", 0),
+            ),
+            "spoke-local-portfolios": dict(
+                puppet_account_id=puppet_account_id,
+                spoke_local_portfolio_name=item_name,
+                product_generation_method=item.get("product_generation_method",
+                                                   constants.PRODUCT_GENERATION_METHOD_DEFAULT),
+                organization=item.get("organization", ""),
+                sharing_mode=item.get("sharing_mode", constants.SHARING_MODE_DEFAULT),
+                associations=item.get("associations", list()),
+                launch_constraints=item.get("constraints", {}).get("launch", []),
+                portfolio=item.get("portfolio"),
+            ),
+            "lambda-invocations": dict(
+                puppet_account_id=puppet_account_id,
+                lambda_invocation_name=item_name,
+                function_name=item.get("function_name"),
+                qualifier=item.get("qualifier", "$LATEST"),
+                invocation_type=item.get("invocation_type"),
+                launch_parameters=item.get("parameters", {}),
+                manifest_parameters=self.get("parameters", {}),
+            ),
+            "code-build-runs": dict(
+                puppet_account_id=puppet_account_id,
+                code_build_run_name=item_name,
+                launch_parameters=item.get("parameters", {}),
+                manifest_parameters=self.get("parameters", {}),
+                project_name=item.get("project_name"),
+                requested_priority=item.get("requested_priority", 0),
+            ),
+            "assertions": dict(
+                puppet_account_id=puppet_account_id,
+                requested_priority=item.get("requested_priority", 0),
+                assertion_name=item_name,
+                expected=item.get("expected"),
+                actual=item.get("actual"),
+            ),
+        }.get(section_name)
 
         tags = item.get(deploy_to).get("tags", [])
         for tag in tags:
@@ -296,10 +339,26 @@ class Manifest(dict):
             regions = tag.get("regions")
             for account in accounts:
                 account_id = str(account.get("account_id"))
-                additional_parameters = dict(
-                    account_id=account_id,
-                    account_parameters=account.get("parameters", {}),
-                )
+                additional_parameters = {
+                    "launches": dict(
+                        account_id=account_id,
+                        account_parameters=account.get("parameters", {}),
+                    ),
+                    "spoke-local-portfolios": dict(
+                        account_id=account_id,
+                    ),
+                    "assertions": dict(
+                        account_id=account_id,
+                    ),
+                    "lambda-invocations": dict(
+                        account_id=account_id,
+                        account_parameters=account.get("parameters", {}),
+                    ),
+                    "code-build-runs": dict(
+                        account_id=account_id,
+                        account_parameters=account.get("parameters", {}),
+                    ),
+                }.get(section_name)
                 if tag_name in account.get("tags"):
                     if isinstance(regions, str):
                         if regions in [
@@ -336,7 +395,7 @@ class Manifest(dict):
 
                         else:
                             raise Exception(
-                                f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {launch_name}"
+                                f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {item_name}"
                             )
                     elif isinstance(regions, list):
                         for region_ in regions:
@@ -359,7 +418,7 @@ class Manifest(dict):
 
                     else:
                         raise Exception(
-                            f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {launch_name}"
+                            f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {item_name}"
                         )
 
         for account_to_deploy_to in item.get(deploy_to).get("accounts", []):
@@ -368,9 +427,26 @@ class Manifest(dict):
 
             account = self.get_account(account_id_of_account_to_deploy_to)
             account_id = account_id_of_account_to_deploy_to
-            additional_parameters = dict(
-                account_id=account_id, account_parameters=account.get("parameters", {}),
-            )
+            additional_parameters = {
+                "launches": dict(
+                    account_id=account_id,
+                    account_parameters=account.get("parameters", {}),
+                ),
+                "spoke-local-portfolios": dict(
+                    account_id=account_id,
+                ),
+                "assertions": dict(
+                    account_id=account_id,
+                ),
+                "lambda-invocations": dict(
+                    account_id=account_id,
+                    account_parameters=account.get("parameters", {}),
+                ),
+                "code-build-runs": dict(
+                    account_id=account_id,
+                    account_parameters=account.get("parameters", {}),
+                ),
+            }.get(section_name)
 
             if isinstance(regions, str):
                 if regions in [
@@ -407,7 +483,7 @@ class Manifest(dict):
 
                 else:
                     raise Exception(
-                        f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {launch_name}"
+                        f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {item_name}"
                     )
             elif isinstance(regions, list):
                 for region_ in regions:
@@ -430,71 +506,68 @@ class Manifest(dict):
 
             else:
                 raise Exception(
-                    f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {launch_name}"
+                    f"Unsupported regions {regions} setting for {constants.LAUNCHES}: {item_name}"
                 )
         return provisioning_tasks
 
-    def get_provisioning_tasks_for_launch_and_region(
-        self, puppet_account_id, launch_name, region
+    def get_tasks_for_launch_and_region(
+            self, puppet_account_id, section_name, launch_name, region
     ):
         return [
             task
-            for task in self.get_tasks_for_launch(puppet_account_id, launch_name)
+            for task in self.get_tasks_for(puppet_account_id, section_name, launch_name)
             if task.get("region") == region
         ]
 
-    def get_provisioning_tasks_for_launch_and_account(
-        self, puppet_account_id, launch_name, account_id
+    def get_tasks_for_launch_and_account(
+            self, puppet_account_id, section_nam, launch_name, account_id
     ):
         return [
             task
-            for task in self.get_tasks_for_launch(puppet_account_id, launch_name)
+            for task in self.get_tasks_for(puppet_account_id, section_nam, launch_name)
             if task.get("account_id") == account_id
         ]
 
-    def get_provisioning_tasks_for_launch_and_account_and_region(
-        self, puppet_account_id, launch_name, account_id, region
+    def get_tasks_for_launch_and_account_and_region(
+            self, puppet_account_id, section_name, launch_name, account_id, region
     ):
         return [
             task
-            for task in self.get_tasks_for_launch(puppet_account_id, launch_name)
+            for task in self.get_tasks_for(puppet_account_id, section_name, launch_name)
             if task.get("account_id") == account_id and task.get("region") == region
         ]
 
     def get_regions_used_for_section_item(
-        self, puppet_account_id, section_name, item_name
+            self, puppet_account_id, section_name, item_name
     ):
-        if section_name == constants.LAUNCHES:
-            return list(
-                set(
-                    task.get("region")
-                    for task in self.get_tasks_for_launch(puppet_account_id, item_name)
-                )
+        return list(
+            set(
+                task.get("region")
+                for task in self.get_tasks_for(puppet_account_id, section_name, item_name)
             )
+        )
 
     def get_account_ids_used_for_section_item(
-        self, puppet_account_id, section_name, item_name
+            self, puppet_account_id, section_name, item_name
     ):
-        if section_name == constants.LAUNCHES:
-            return list(
-                set(
-                    task.get("account_id")
-                    for task in self.get_tasks_for_launch(puppet_account_id, item_name)
-                )
+        return list(
+            set(
+                task.get("account_id")
+                for task in self.get_tasks_for(puppet_account_id, section_name, item_name)
             )
+        )
 
     def get_account_ids_and_regions_used_for_section_item(
-        self, puppet_account_id, section_name, item_name
+            self, puppet_account_id, section_name, item_name
     ):
-        if section_name == constants.LAUNCHES:
-            result = dict()
-            for task in self.get_tasks_for_launch(puppet_account_id, item_name):
-                if result.get(task.get("account_id")) is None:
-                    result[task.get("account_id")] = list()
-                result[task.get("account_id")].append(task.get("region"))
-            for account_id in result.keys():
-                result[account_id] = list(set(result[account_id]))
-            return result
+        result = dict()
+        for task in self.get_tasks_for(puppet_account_id, section_name, item_name):
+            if result.get(task.get("account_id")) is None:
+                result[task.get("account_id")] = list()
+            result[task.get("account_id")].append(task.get("region"))
+        for account_id in result.keys():
+            result[account_id] = list(set(result[account_id]))
+        return result
 
     def get_mapping(self, mapping, account_id, region):
         manifest_mappings = self.get("mappings")
@@ -528,7 +601,7 @@ class Manifest(dict):
         return result
 
     def get_actions_from(
-        self, launch_name, pre_or_post, launch_or_spoke_local_portfolio
+            self, launch_name, pre_or_post, launch_or_spoke_local_portfolio
     ):
         logger.info(
             f"get_actions_from for {launch_or_spoke_local_portfolio}.{launch_name}"
@@ -563,9 +636,9 @@ class Manifest(dict):
             account_regions.append(account.get("default_region"))
 
             enabled_regions = (
-                account.get("enabled", [])
-                + account.get("regions_enabled", [])
-                + account.get("enabled_regions", [])
+                    account.get("enabled", [])
+                    + account.get("regions_enabled", [])
+                    + account.get("enabled_regions", [])
             )
             if len(enabled_regions) == 0:
                 raise Exception(
@@ -579,8 +652,8 @@ class Manifest(dict):
                 if account.get("organization", "") != "":
                     organization = account.get("organization")
                     if (
-                        organization
-                        not in sharing_policies_by_region[r]["organizations"]
+                            organization
+                            not in sharing_policies_by_region[r]["organizations"]
                     ):
                         sharing_policies_by_region[r]["organizations"].append(
                             organization
@@ -630,11 +703,11 @@ class Manifest(dict):
         return accounts_by_region
 
     def get_task_defs_from_details(
-        self,
-        puppet_account_id,
-        launch_name,
-        configuration,
-        launch_or_spoke_local_portfolio,
+            self,
+            puppet_account_id,
+            launch_name,
+            configuration,
+            launch_or_spoke_local_portfolio,
     ):
         launch_details = self.get(launch_or_spoke_local_portfolio).get(launch_name)
         accounts = self.get("accounts")
@@ -776,10 +849,10 @@ def convert_to_graph(expanded_manifest, G):
     for section in sections:
         for item_name, item_details in expanded_manifest.get(section, {}).items():
             uid = f"{section}|{item_name}"
-            data = dict(section=section, item_name=item_name,)
+            data = dict(section=section, item_name=item_name, )
             data.update(item_details)
             G.add_nodes_from(
-                [(uid, data),]
+                [(uid, data), ]
             )
 
     mapping = dict()
